Add inhibit-read-only flag
The buffer-read-only flag is set to true at the end of this
function. Which means that running the same function again with an
open buffer will produce an error.

This fix sets inhibit-read-only to true in a let expression, which
means that calling the function again will simply overwrite the
appropriate buffer.